+++ src/main/java/com/es2/vadebicicleta/es2/vadebicicleta/aluguel/domain/dto/AluguelDevolucaoDTO.java
@@ -20,7 +20,7 @@ public class AluguelDevolucaoDTO {
     @NotNull
     private String horaFim;
     @NotNull
-    private BigDecimal cobranca;
+    private Long cobranca;
     @NotNull
     private Integer ciclista;
 }
