+++ src/main/java/com/es2/vadebicicleta/es2/vadebicicleta/aluguel/domain/dto/AluguelOutDTO.java
@@ -15,7 +15,7 @@ public class AluguelOutDTO {
     private String horaInicio;
     private Integer trancaFim;
     private String horaFim;
-    private BigDecimal cobranca;
+    private Long cobranca;
     private Integer ciclista;
     private Integer trancaInicio;
 }
